requirements.txt
langchain
langchain-google-genai
google-generativeai
supabase
python-dotenv
tiktoken
requests
